Issue-01  solved.I have comment out it.I will add user in next update

+++ src/main/java/com/fahim/shoppingcard/request/UpdateProductRequest.java
@@ -3,12 +3,13 @@ package com.fahim.shoppingcard.request;
 import com.fahim.shoppingcard.model.Category;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 @Data
 public class UpdateProductRequest {
     private long id;
     private String name;
-    private BigInteger price;
+    private BigDecimal price;
     private String brand;
     private int inventory;
     private String description;
